Move the helpers to a standalone file, include them in Spree::BaseMailer.
